requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cors==1.0.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.12
fastjsonschema==2.20.0
filelock==3.10.7
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
future==1.0.0
gast==0.6.0
gevent==24.11.1
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-cloud-vision==3.4.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==3.2.0
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.4
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.0.2
opencv-python==4.7.0.72
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==23.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Pillow==9.4.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.22.2
protobuf==4.22.1
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2023.3.23
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.11.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
textblob==0.17.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tldextract==5.2.0
tomli==2.0.2
torch==2.0.0
torchvision==0.15.1
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
zope.event==5.0
zope.interface==7.2
